backend/requirements.txt
# FastAPI
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.17

# Database
sqlalchemy[asyncio]>=2.0.36
asyncpg>=0.30.0
alembic>=1.14.0

# Validation & Settings
pydantic>=2.10.0
pydantic-settings>=2.6.0
email-validator>=2.0.0

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Redis
redis>=5.2.0

# AWS
boto3>=1.35.0

# AI
google-generativeai>=0.8.0

# HTTP Client
httpx>=0.28.0

# Utils
python-dotenv>=1.0.1

# Development
ruff>=0.8.0
pytest>=8.3.0
pytest-asyncio>=0.24.0
